prueba #4 fix2.0 del admin panel

+++ db/models/producto.py
@@ -4,7 +4,7 @@ from typing import Optional
 class Product(BaseModel):
     id : str | None = None
     name : str
-    precio : int
+    precio : float
     categoria : str
     imagen : str
     descripcion : str
